refactor: view, domain 분리
