+++ src/main/kotlin/racingcar/Main.kt
@@ -1,5 +1,10 @@
 package racingcar
 
+import racingcar.domain.RacingGame
+import racingcar.domain.RandomGenerator
+import racingcar.view.InputView
+import racingcar.view.ResultView
+
 fun main() {
     val names = InputView.inputCarNames()
     val tryCount = InputView.inputTryCount()
