+++ src/main/kotlin/racingcar/domain/Positions.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 data class Positions(val values: List<Position>) {
     constructor(vararg values: Position): this(values.toList())
