+++ src/main/kotlin/racingcar/domain/RacingGame.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 class RacingGame(
     randomGenerator: RandomGenerator,
