+++ src/main/kotlin/racingcar/domain/RandomGenerator.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 fun interface RandomGenerator {
